filenames to match google guide, clang-format run one last time and slide_test run to double check it's all good still.
