+++ src/CMakeLists.txt
@@ -11,7 +11,7 @@ if(CAIRO_FOUND)
 	message(STATUS \t${CAIRO_LIBRARIES})
 endif()
 
-add_library(slide-lib Slide.cpp Color.cpp WCairo.cpp Documents.cpp Encoding.cpp)
+add_library(slide-lib slide.cc color.cc wrapped_cairo.cc documents.cc encoding.cc)
 target_link_libraries(slide-lib)
 target_link_libraries(slide-lib INTERFACE ${CAIRO_LIBRARIES})
 
@@ -19,7 +19,7 @@ add_library(slide-core INTERFACE)
 target_include_directories(slide-core INTERFACE ${CMAKE_CURRENT_SOURCE_DIR} ${CAIRO_INCLUDE_DIRS})
 target_link_libraries(slide-core INTERFACE slide-lib)
 
-add_library(app App.cpp Color.cpp)
+add_library(app app.cc color.cc)
 target_include_directories(app INTERFACE ${CMAKE_CURRENT_SOURCE_DIR} ${CAIRO_INCLUDE_DIRS})
 target_link_libraries(app PRIVATE webview slide-core)
 
@@ -48,12 +48,12 @@ add_library(json INTERFACE)
 target_include_directories(json INTERFACE ${CMAKE_CURRENT_SOURCE_DIR} ${CAIRO_INCLUDE_DIRS})
 
 if(WIN32)
-	add_executable(slide WIN32 main.cpp ${CMAKE_SOURCE_DIR}/res/main.rc)
+	add_executable(slide WIN32 main.cc ${CMAKE_SOURCE_DIR}/res/main.rc)
 elseif(APPLE)
-	add_executable(slide main.cpp ${CMAKE_SOURCE_DIR}/res/main.rc)
-	set_source_files_properties(main.cpp PROPERTIES COMPILE_FLAGS "-x objective-c++")
+	add_executable(slide main.cc ${CMAKE_SOURCE_DIR}/res/main.rc)
+	set_source_files_properties(main.cc PROPERTIES COMPILE_FLAGS "-x objective-c++")
 else()
-	add_executable(slide main.cpp ${CMAKE_SOURCE_DIR}/res/main.rc)
+	add_executable(slide main.cc ${CMAKE_SOURCE_DIR}/res/main.rc)
 endif()
 target_link_libraries(slide PRIVATE slide-core webview json app slide-lib)
