+++ src/color.cc
@@ -2,4 +2,4 @@
 // Created by Patrick Mintram on 28/11/2017.
 //
 
-#include "Color.h"
+#include "color.h"
